Feature and Admin Upgrade(cpp and sql)
Features and Admin Controls added:
- Groups now work.  Difficulty is now balanced across group size
- Debuff added for groups  trying to exceed the difficulty modifier
- Debuff can be enabled and disabled
- Spellpower is now scaled based on character level and modifier
- Stat modifier can now be changed in config to reel in melee damage
- Individual modification of each dungeon/raid difficulty
- Max level cap to keep high level characters from getting buffed
- Max level cap modifiable in config for each dungeon/raid
- Dungeon/Raid types still used as a catch all if not found as an
individual instance in the config

A database table is included to manage groups and the difficulty
modifier that each group member has applied to them.  It also
deals with server restarts and players having to log out and  log
back in to try to correct stats.

Updated README.md file with current instructions and typo fixes

Fixes #9